+++ glue/src/stateful/bootstrap/wire.rs
@@ -4,30 +4,30 @@ use commonware_consensus::{marshal::core::Variant, simplex::types::Finalization}
 use commonware_cryptography::certificate::Scheme;
 
 pub(crate) enum Tag {
-    RequestLatest,
-    Finalization,
+    Request,
+    Response,
 }
 
 impl Tag {
     pub(crate) fn read(reader: &mut impl Buf) -> Result<Self, Error> {
         match u8::read(reader)? {
-            0 => Ok(Self::RequestLatest),
-            1 => Ok(Self::Finalization),
+            0 => Ok(Self::Request),
+            1 => Ok(Self::Response),
             n => Err(Error::InvalidEnum(n)),
         }
     }
 }
 
-/// A message exchanged with peers over the p2p channel during floor discovery.
+/// A message exchanged with peers over the p2p channel during bootstrap.
 pub(crate) enum Message<S, V>
 where
     S: Scheme,
     V: Variant,
 {
     /// Request the receiver's latest [`Finalization`].
-    RequestLatest,
-    /// A [`Finalization`], sent in response to a [`Message::RequestLatest`].
-    Finalization(Finalization<S, V::Commitment>),
+    Request,
+    /// A [`Finalization`], sent in response to a [`Message::Request`].
+    Response(Finalization<S, V::Commitment>),
 }
 
 impl<S, V> Write for Message<S, V>
@@ -37,10 +37,10 @@ where
 {
     fn write(&self, writer: &mut impl BufMut) {
         match self {
-            Self::RequestLatest => {
+            Self::Request => {
                 0u8.write(writer);
             }
-            Self::Finalization(finalization) => {
+            Self::Response(finalization) => {
                 1u8.write(writer);
                 finalization.write(writer);
             }
@@ -55,8 +55,8 @@ where
 {
     fn encode_size(&self) -> usize {
         1 + match self {
-            Self::RequestLatest => 0,
-            Self::Finalization(finalization) => finalization.encode_size(),
+            Self::Request => 0,
+            Self::Response(finalization) => finalization.encode_size(),
         }
     }
 }
