+++ glue/src/stateful/tests/multi_db_app.rs
@@ -5,11 +5,11 @@ use crate::{
         reporter::MonitorReporter,
     },
     stateful::{
+        bootstrap::{Bootstrap, Config as BootstrapConfig},
         db::{
             p2p::standard as qmdb_resolver, DatabaseSet, Merkleized as _, SyncEngineConfig,
             Unmerkleized as _,
         },
-        floor_discovery::{Config as FloorDiscoveryConfig, FloorDiscovery},
         Application, Config as StatefulConfig, Proposed, Stateful as StatefulActor, SyncPlan,
     },
 };
@@ -386,7 +386,7 @@ impl EngineDefinition for MultiDbEngine {
             (3, TEST_QUOTA), // backfill
             (4, TEST_QUOTA), // broadcast
             (5, TEST_QUOTA), // qmdb sync resolvers (muxed)
-            (6, TEST_QUOTA), // floor discovery
+            (6, TEST_QUOTA), // bootstrap
         ]
     }
 
@@ -452,7 +452,7 @@ impl EngineDefinition for MultiDbEngine {
         let backfill_network = channels.next().unwrap();
         let broadcast_network = channels.next().unwrap();
         let qmdb_resolver_network = channels.next().unwrap();
-        let floor_discovery_network = channels.next().unwrap();
+        let bootstrap_network = channels.next().unwrap();
 
         // Mux the QMDB resolver channel into two subchannels (one per database).
         let (mux, mut mux_handle) = Muxer::new(
@@ -527,21 +527,20 @@ impl EngineDefinition for MultiDbEngine {
         let should_state_sync = plan.should_state_sync(self.enable_state_sync && delayed);
         let provider = ConstantProvider::new(scheme.clone());
 
-        let (floor_discovery, floor_discovery_mailbox) =
-            FloorDiscovery::new(FloorDiscoveryConfig {
-                context: context.child("floor_discovery"),
-                provider: provider.clone(),
-                strategy: Sequential,
-                capacity: NZUsize!(100),
-                blocker: oracle.control(public_key.clone()),
-                retry_timeout: NZDuration!(Duration::from_millis(100)),
-            });
-        floor_discovery.start(floor_discovery_network);
+        let (bootstrap, bootstrap_mailbox) = Bootstrap::new(BootstrapConfig {
+            context: context.child("bootstrap"),
+            provider: provider.clone(),
+            strategy: Sequential,
+            capacity: NZUsize!(100),
+            blocker: oracle.control(public_key.clone()),
+            retry_timeout: NZDuration!(Duration::from_millis(100)),
+        });
+        bootstrap.start(bootstrap_network);
         let mut state_sync_height = if should_state_sync {
-            let finalization = floor_discovery_mailbox
+            let finalization = bootstrap_mailbox
                 .subscribe()
                 .await
-                .expect("floor discovery stopped");
+                .expect("bootstrap stopped");
             plan = plan.with_floor(finalization);
             None
         } else {
@@ -647,9 +646,9 @@ impl EngineDefinition for MultiDbEngine {
         // Start marshal actor with monitored reporters.
         marshal_actor.start(marshal_reporters, buffer, resolver);
 
-        // Attach the marshal to floor discovery, transitioning it to serving. A syncing node has
-        // already consumed its floor above; a source attaches without ever soliciting peers.
-        floor_discovery_mailbox.attach(marshal_mailbox.clone());
+        // Attach the marshal to bootstrap. A syncing node has already consumed its floor above; a
+        // source attaches without ever soliciting peers.
+        bootstrap_mailbox.attach(marshal_mailbox.clone());
 
         if should_state_sync {
             let finalization = sync_floor.expect("sync floor missing");
