+++ glue/src/stateful/tests/single_db_app.rs
@@ -5,11 +5,11 @@ use crate::{
         reporter::MonitorReporter,
     },
     stateful::{
+        bootstrap::{Bootstrap, Config as BootstrapConfig},
         db::{
             p2p::standard as qmdb_resolver, DatabaseSet, Merkleized as _, SyncEngineConfig,
             Unmerkleized as _,
         },
-        floor_discovery::{Config as FloorDiscoveryConfig, FloorDiscovery},
         Application, Config as StatefulConfig, Proposed, Stateful as StatefulActor, SyncPlan,
     },
 };
@@ -326,7 +326,7 @@ impl EngineDefinition for SingleDbEngine {
             (3, TEST_QUOTA), // backfill
             (4, TEST_QUOTA), // broadcast
             (5, TEST_QUOTA), // qmdb sync resolver
-            (6, TEST_QUOTA), // floor discovery
+            (6, TEST_QUOTA), // bootstrap
         ]
     }
 
@@ -374,7 +374,7 @@ impl EngineDefinition for SingleDbEngine {
         let backfill_network = channels.next().unwrap();
         let broadcast_network = channels.next().unwrap();
         let qmdb_resolver_network = channels.next().unwrap();
-        let floor_discovery_network = channels.next().unwrap();
+        let bootstrap_network = channels.next().unwrap();
 
         // Marshal resolver
         let resolver_cfg = marshal_resolver::Config {
@@ -437,21 +437,20 @@ impl EngineDefinition for SingleDbEngine {
         let should_state_sync = plan.should_state_sync(self.enable_state_sync && delayed);
         let provider = ConstantProvider::new(scheme.clone());
 
-        let (floor_discovery, floor_discovery_mailbox) =
-            FloorDiscovery::new(FloorDiscoveryConfig {
-                context: context.child("floor_discovery"),
-                provider: provider.clone(),
-                strategy: Sequential,
-                capacity: NZUsize!(100),
-                blocker: oracle.control(public_key.clone()),
-                retry_timeout: NZDuration!(Duration::from_millis(100)),
-            });
-        floor_discovery.start(floor_discovery_network);
+        let (bootstrap, bootstrap_mailbox) = Bootstrap::new(BootstrapConfig {
+            context: context.child("bootstrap"),
+            provider: provider.clone(),
+            strategy: Sequential,
+            capacity: NZUsize!(100),
+            blocker: oracle.control(public_key.clone()),
+            retry_timeout: NZDuration!(Duration::from_millis(100)),
+        });
+        bootstrap.start(bootstrap_network);
         let mut state_sync_height = if should_state_sync {
-            let finalization = floor_discovery_mailbox
+            let finalization = bootstrap_mailbox
                 .subscribe()
                 .await
-                .expect("floor discovery stopped");
+                .expect("bootstrap stopped");
             plan = plan.with_floor(finalization);
             None
         } else {
@@ -538,9 +537,9 @@ impl EngineDefinition for SingleDbEngine {
         // Start marshal actor with monitored reporters.
         marshal_actor.start(marshal_reporters, buffer, resolver);
 
-        // Attach the marshal to floor discovery, transitioning it to serving. A syncing node has
-        // already consumed its floor above; a source attaches without ever soliciting peers.
-        floor_discovery_mailbox.attach(marshal_mailbox.clone());
+        // Attach the marshal to bootstrap. A syncing node has already consumed its floor above; a
+        // source attaches without ever soliciting peers.
+        bootstrap_mailbox.attach(marshal_mailbox.clone());
 
         if should_state_sync {
             let finalization = sync_floor.expect("sync floor missing");
